Frontend/requirements.txt
Django==3.2.9
requests==2.26.0
python-dotenv==0.19.2